Webappconfig/{config.py,worker.py}: Adds configurable --copy option.
This allows for altering the value of vhost_link_type in
/etc/vhosts/webapp-config to copy files by default when installing
a webapp as opposed to soft-linking or hard-linking.  This patch also
repairs the previous damage in which the behavior of the --copy flag
was being ignored; due to the fact that the boolean g_copy was not
being passed down throughout the program because of the extra --cp
flag.  Man page alterations have also been made to reflect the removal
of the --cp flag in the program.

X-Gentoo-Bug: 231482
X-Gentoo-Bug-URL: https://bugs.gentoo.org/231482

Signed-off-by: Anthony G. Basile <blueness@gentoo.org>