WebappConfig/permissions.py: Corrected syntax for eval() function.
When the permissions octal was being passed to the eval() function they
were previously not being evaluated with python3.2 due to the fact that
the octal value was not being declared as an octal. To fix this, it was
simply stated that the octal value was one by appending '0o' to the
values being piped into the eval() function.

X-GENTOO-BUG: 481298
X-GENTOO-BUG-URL: https://bugs.gentoo.org/481298
Signed-off-by: Anthony G. Basile <blueness@gentoo.org>