WebappConfig/compat.py: Adds compatibility check for file_md5().
When the file_md5() function is called in python3.2+ a hash file is
supposed to be created, and on a if an object is unicode, it needs to be
encoded prior to hashing. In order to do this, a file called compat.py
has a function in it called create_md5(); this function checks the
hexidecimal python version and based on the version, either encodes the
file before it hashes, or simply returns the md5 hash.

Signed-off-by: Anthony G. Basile <blueness@gentoo.org>