WebappConfig/compat.py: Revamps create_md5() function.
This second revamp changes the way in which python is opening
the file to create an md5 hash of. Upon calling the open()
function it opens the file to read it, or as a binary. This
allows for consistent behavior when trying to open either a
binary file, or a text file to hash. This method is also much
cleaner and sufficient for forward and backwards compatibility.

Conflicts:
	WebappConfig/compat.py