WebappConfig/config.py: Modifies default interpolation style.
Due to python3.2 using one interpolation style when running ConfigParser
it was necessary to change the interpolation style to the Extended
Interpolation style which allows ConfigParser to interpolate such
variables that follow the style of ${} in config.py and the config
file in /etc/vhosts/webapp-config. For total backwards compatibility
we also needed to change the interpolation style being used in config.py
because it was using a Basic Interpolation style of %(variable)s.
This change has been tested with both python3.2 and python2.7.