WebappConfig/config.py: Prevents overwriting config variables.
When setting vhost_subdomain_# variables in /etc/vhosts/webapp-config
they have a possibility of being overwritten by the split_hostname()
function. To see this happen you should set vhost_subdomain_1 to a
value and try and use it as a variable in your config file. This will
then get overwritten by your vhost_hostname. This commit will prevent
that from occuring while also having the get() method ignore the
inexistence of a vhost_subdomain_# variable in your config file,
due to the fact that if the variable does exist it will be used, and
if the variable does not exist then it will get overwritten by a
piece of vhost_hostname that is separated by the "." delimiter.

X-Gentoo-Bug: 300250
X-Gentoo-Bug-URL: https://bugs.gentoo.org/300250
X-Gentoo-Bug: 349491
X-Gentoo-Bug-URL: https://bugs.gentoo.org/349491