WebappConfig/config.py: Uses maybe_get() function instead of get().
This changes the use of the get() function in the split_hostname()
function in favor of the maybe_get() function. This change still
allows vhost_subdomain_# variables to be set in /etc/vhosts/webapp-config
without the risk of it being overwritten except due to the use of the
maybe_get() function it will no longer give warnings to the user if
a vhost_subdomain_# variable is not found. This is ideal because
it removes the need of masking the warning in the get() function.

X-Gentoo-Bug: 300250
X-Gentoo-Bug-URL: https://bugs.gentoo.org/300250
X-Gentoo-Bug: 349491
X-Gentoo-Bug-URL: https://bugs.gentoo.org/349491