WebappConfig/ebuild.py: Alters doctest output to default values.
On line 232 on ebuild.py the testing of the documentation was failing
due to not returning full variables and instead returning a varaible
along with "..." in the place of any part of a variable that can change.
The placement of the "..." was causing the doctest to fail upon emerge.

X-Gentoo-Bug: 430010
X-Gentoo-Bug-URL: https://bugs.gentoo.org/430010