WebappConfig/ebuild.py: Nulls doctest code in run_var()
Due to the variable nature of the output of the run_var() function it
is unrealistic to create doctest code that will not fail a test. This
is mainly due to the fact that there are particular variables that may
never be the same on any two user's systems. It has been decided that
the best solution would be to simply "comment out" the doctest code
to prevent it from running and causing failures.

X-Gentoo-Bug: 430010
X-Gentoo-Bug-URL: https://bugs.gentoo.org/430010