doc/webapp-config.8.xml: Improves man page synopsis.
Improvements done to the synopsis of webapp-config include the
following: including the --copy, --soft, and --secure flags with the
synopsis for the -U flag to indicate to the user that these flags are
also usable with this flag. Adding the necessary opts app-name and
app-version for the -C flag to indicate to users that these two args are
necessary when uninstalling a webapp. And adding the --secure flag to
the synopsis for the -C flag to indicate that it is an optional arg if
the --secure flag is used with the -I flag or -U flag.

X-Gentoo-Bug: 496568
X-Gentoo-Bug-URL: https://bugs.gentoo.org/496568