WebappConfig/db.py: Slightly alters prune_database() algorithm.
When the prune_database() function checks to see if a webapp is
installed in the directory the database says it's supposed to be
installed in it originally checked to see if a .webapp file existed
in the directory. But due to the fact that a .webapp file could exist
in a directory and not be the webapp that is listed as installed in
that directory it was decided that checking the existance of the
.webapp-<webapp>-<version> file would be better as it confirms that
a webapp exists in that directory and it is the webapp that it is
supposed to be.