WebappConfig/db.py: Properly closes install file after writing.
When webapp-config cleans out a webapp installation it needs to
remove the database entry of the file from it's designated installs
file. If more than one database entry is in the installs file it
should still list the database entry for the webapp that isn't being
removed. The logic is all there in db.py but after the file has
rewritten the entries it doesn't close the file properly and causes
a failure in the check right after the write(), which then unlinks
the installs file and removes the entries from webapp-config's
database.

X-Gentoo-Bug:494520
X-Gentoo-Bug-URL: https://bugs.gentoo.org/494520