webapp-config.8.xml: clarifies man page for --dd and --vf flags
This commit corrects a typo where the short option for --default-dirs
is --vd instead of --dd, this has been corrected. Secondly, it also
attempts to make it clearer to the user that the --dd and --vf options
are two separate options and should not be treated equally by listing
them in two separate variable list entries.

X-Gentoo-Bug: 257592
X-Gentoo-Bug-URL: https://bugs.gentoo.org/257592