config.py: Fixes package version checking regression
Previously webapp-config would not do any sanity checks when
setting the package version. After adding a sanity check in 1.54
I made the mistake of not being flexible enough and this caused
a regression that prevented web apps with versions such as
"20140929d"[1] to be installed. This commit fixes that while still
allowing for some sanity checking.

[1]: https://github.com/gentoo/webapp-config/issues/2